Refactor: move directories
