+++ app/login/client.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
 import { useSessionId } from '../../context/SessionIdContext';
-import { apiLogin } from '@/lib/api/login';
+import { apiLogin } from '@/lib/api/calls/login';
 
 import { useAnimationTimeout } from '@/lib/hooks/timeout';
