+++ app/mypage/change-pw/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { apiProfileUpdatePW } from '@/lib/api/mypage';
+import { apiProfileUpdatePW } from '@/lib/api/calls/mypage';
 import { useState } from 'react';
 
 import { useRouter } from 'next/navigation';
