+++ app/register/steps/step1.tsx
@@ -11,7 +11,7 @@ import { HStack, VStack } from '@/components/basic/stack';
 import { useAnimationTimeout } from '@/lib/hooks/timeout';
 
 import { RightIcon } from '@/icons';
-import { apiSendEmail } from '@/lib/api/login';
+import { apiSendEmail } from '@/lib/api/calls/login';
 import Spinner from '@/components/basic/spinner';
 
 const validateEmail = (email: string) => {
